[CONTROLLER/RECORDER] splits big file
