+++ server/controller/recorder/cache/dhcp_port.go
@@ -1,0 +1,59 @@
+/**
+ * Copyright (c) 2023 Yunshan Networks
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cache
+
+import (
+	cloudmodel "github.com/deepflowio/deepflow/server/controller/cloud/model"
+	"github.com/deepflowio/deepflow/server/controller/db/mysql"
+	. "github.com/deepflowio/deepflow/server/controller/recorder/common"
+)
+
+func (b *DiffBaseDataSet) addDHCPPort(dbItem *mysql.DHCPPort, seq int, toolDataSet *ToolDataSet) {
+	vpcLcuuid, _ := toolDataSet.GetVPCLcuuidByID(dbItem.VPCID)
+	b.DHCPPorts[dbItem.Lcuuid] = &DHCPPort{
+		DiffBase: DiffBase{
+			Sequence: seq,
+			Lcuuid:   dbItem.Lcuuid,
+		},
+		Name:         dbItem.Name,
+		RegionLcuuid: dbItem.Region,
+		AZLcuuid:     dbItem.AZ,
+		VPCLcuuid:    vpcLcuuid,
+	}
+	b.GetLogFunc()(addDiffBase(RESOURCE_TYPE_DHCP_PORT_EN, b.DHCPPorts[dbItem.Lcuuid]))
+}
+
+func (b *DiffBaseDataSet) deleteDHCPPort(lcuuid string) {
+	delete(b.DHCPPorts, lcuuid)
+	log.Info(deleteDiffBase(RESOURCE_TYPE_DHCP_PORT_EN, lcuuid))
+}
+
+type DHCPPort struct {
+	DiffBase
+	Name         string `json:"name"`
+	RegionLcuuid string `json:"region_lcuuid"`
+	AZLcuuid     string `json:"az_lcuuid"`
+	VPCLcuuid    string `json:"vpc_lcuuid"`
+}
+
+func (d *DHCPPort) Update(cloudItem *cloudmodel.DHCPPort) {
+	d.Name = cloudItem.Name
+	d.RegionLcuuid = cloudItem.RegionLcuuid
+	d.AZLcuuid = cloudItem.AZLcuuid
+	d.VPCLcuuid = cloudItem.VPCLcuuid
+	log.Info(updateDiffBase(RESOURCE_TYPE_DHCP_PORT_EN, d))
+}
