+++ server/controller/recorder/cache/pod_cluster.go
@@ -1,0 +1,61 @@
+/**
+ * Copyright (c) 2023 Yunshan Networks
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cache
+
+import (
+	cloudmodel "github.com/deepflowio/deepflow/server/controller/cloud/model"
+	"github.com/deepflowio/deepflow/server/controller/db/mysql"
+	. "github.com/deepflowio/deepflow/server/controller/recorder/common"
+)
+
+func (b *DiffBaseDataSet) addPodCluster(dbItem *mysql.PodCluster, seq int) {
+	b.PodClusters[dbItem.Lcuuid] = &PodCluster{
+		DiffBase: DiffBase{
+			Sequence: seq,
+			Lcuuid:   dbItem.Lcuuid,
+		},
+		Name:            dbItem.Name,
+		ClusterName:     dbItem.ClusterName,
+		RegionLcuuid:    dbItem.Region,
+		AZLcuuid:        dbItem.AZ,
+		SubDomainLcuuid: dbItem.SubDomain,
+	}
+	b.GetLogFunc()(addDiffBase(RESOURCE_TYPE_POD_CLUSTER_EN, b.PodClusters[dbItem.Lcuuid]))
+}
+
+func (b *DiffBaseDataSet) deletePodCluster(lcuuid string) {
+	delete(b.PodClusters, lcuuid)
+	log.Info(deleteDiffBase(RESOURCE_TYPE_POD_CLUSTER_EN, lcuuid))
+}
+
+type PodCluster struct {
+	DiffBase
+	Name            string `json:"name"`
+	ClusterName     string `json:"cluster_name"`
+	RegionLcuuid    string `json:"region_lcuuid"`
+	AZLcuuid        string `json:"az_lcuuid"`
+	SubDomainLcuuid string `json:"sub_domain_lcuuid"`
+}
+
+func (p *PodCluster) Update(cloudItem *cloudmodel.PodCluster) {
+	p.Name = cloudItem.Name
+	p.ClusterName = cloudItem.ClusterName
+	p.RegionLcuuid = cloudItem.RegionLcuuid
+	p.AZLcuuid = cloudItem.AZLcuuid
+	p.SubDomainLcuuid = cloudItem.SubDomainLcuuid
+	log.Info(updateDiffBase(RESOURCE_TYPE_POD_CLUSTER_EN, p))
+}
